Chart added on Stats tab that shows practice minutes over last 7 days. Chart recognizes when data is present. Basic stylization. Formatted within tableViewCell with inset.
Added an update to the fetch results controller "didChange" methods. Changing the date of a duplicate practice session caused a crash because of section/session ordering. A delete and insert was required to "move" the date to its new position.

Added two sort descriptors (session and section sort). Section sort being first since it's used to display titles for sections and needs the right key path.